Function to obtain a detailed visualization of specific results of functional relevance analysis

The function highlight the links between defined genes and/or pathways to better inspect the results obtained with `gene_functional_relevance()` analysis

Usage

funct_rel_heatmap(
  pct,
  adj,
  target_g = NULL,
  target_ptw = NULL,
  weight,
  n_cores = 1,
  file_name = NULL,
  colors = NULL,
  width = 200,
  height = 200,
  res = 300,
  units = "mm",
  ...
)

Arguments

pct

output of the `pathway_cross_talk()` function

adj

adjacency matrix used for PCT calculation

target_g

vector with the genes of interest. If provided, the function will consider only the genes that interact with these in the plot

target_ptw

vector with the pathways of interest. If provided, the function will consider only the CT where these pathways are involved

weight

weight of the genes. If provided, the function will plot the product of the weights in the heatmap. If `NULL` the function will lot only the presence/absence of the links (1-0)

n_cores

number of cores to be used by the function

file_name

file name of the plot. If not specified, it will be saved as "funct_rel_heatmap.jpeg"

colors

vector of two colors to be used in the heatmap. If not provided, the functions used "blue" and "red"

width, height, res, units

graphical value of `jpeg()` function

...

further arguments passed to `Heatmap` function

Details

The function uses the same inputs used for `gene_functional_relevance()` to retrieve the gene interactions of specific `target_g` and/or `target_ptw`. The function produces an heatmap with the genes that interacts with the user-provided ones for each pathway considered. In the heatmap can be reported the product of the weight of the genes if `weight` is provided, or the existence of the link (0-1) otherwise.