Enrichment map

Enrichment map

Usage

enrichment_map(
  x,
  gs_list,
  method = c("overlap", "jaccard"),
  coeff = NULL,
  all_gs = TRUE,
  weight.within = 4,
  weight.between = 1,
  file.prefix = "en_map",
  comm_method = "fastgreedy",
  vertex.label.cex = 1,
  vertex.label.dist = 0.5,
  vertex.label.degree = pi/2,
  vertex.label.font = 2,
  gs_list_size = NULL,
  set_sim_df = NULL,
  plot_flag = TRUE,
  vertex.color = NULL,
  img.width = 200,
  img.height = 200,
  img.res = 300,
  min_comm_size = 2,
  vertex.size.min = 2,
  vertex.size.max = 5,
  edge.wd.min = 1,
  edge.wd.max = 4,
  mark.groups = T,
  vertex.shape = "circle",
  vertex.pie = NULL,
  vertex.color.pal = NULL,
  n_width = 3,
  n_size = 3,
  L = NULL,
  score.decreasing = TRUE,
  n_name = 1,
  legend.cex = 0.7,
  ...
)

Arguments

x

named vector of pathway scores

gs_list

gene set list

method

'overlap' or 'jaccard'

coeff

threshold for the similarity score between two gene sets

all_gs

TRUE/FALSE

weight.within

value to weight the attraction between two vertices of the same community

weight.between

value to weight the attraction between two vertices of two disntict communities

comm_method

one among 'auto', 'fastgreedy' and 'multilev'

vertex.label.cex

vertex label size

vertex.label.dist

vertex label distance from veritex

vertex.label.degree

vertex label orientation

vertex.label.font

vertex label font

gs_list_size

gene set list size

set_sim_df

optional data frame with three columns 'set1', 'set2' and 'sim'

plot_flag

TRUE/FALSE

vertex.color

vertex color

img.width

image width

img.height

image height

img.res

image resolution

min_comm_size

min community size

vertex.size.min

min vertex size

vertex.size.max

max vertex size

edge.wd.min

min edge widht

edge.wd.max

max edge width

mark.groups

TRUE/FALSE

vertex.shape

vertex shape

vertex.pie

vertex pie

...

additional parameters of plot.igraph

file

output file

Value

list of two data frames containing pathway network and vertex similarity